drm/amdgpu: Add DM (Display Manager)
This is the interface we use to bridge the drm modesetting APIs
with the internal DAL APIs.  The idea behind this is to allow
us to separate the hw abstractions and programming sequences
and the high level APIs.  It also allows us to easily change
internal APIs without affecting the higher level APIs and
allows us to easily support changes to the high level APIs such
as adding plane support and transitioning to the atomic
modesetting API.

Signed-off-by: Harry Wentland <harry.wentland@amd.com>