drm/amd/dal: remove sink from MST link
When we remove MST display downstream, we should remove
sink during connector destruction.

Also do not report error if sink already removed from link.
This can happen in case of physical disconnect of mst branch.

Links will be removed in dc_link_detect. But fake removal in
MST connector destruction is not of harm, as sink not in
array already

Signed-off-by: Mykola Lysenko <Mykola.Lysenko@amd.com>
Acked-by: Jordan Lazare <Jordan.Lazare@amd.com>