drm/amd/dal: fix headless hotplug for MST
This mirrors the fix to handle the non-MST case for xinit
when there's no desktop manager calling mode set for us.

This will allow MST to work in the case of going to headless
state then plug back in the same configuration (be it one MST
capable display or a daisy chain), which was broken by the
xinit fix before. This will also allow MST displays to light
up from hotplug in xinit environment, which was not supported
before.

Signed-off-by: Eric Yang <eric.yang2@amd.com>
Acked-by: Harry Wentland <harry.wentland@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>