drm/radeon: add new PX quirk to force APTX power control
ATPX is the ACPI method for controlling AMD PowerXpress laptops.
There are flags to indicate which methods are supported.  If
the dGPU power down flag is not supported, the driver needs to
implement the dGPU power down manually.  We had previously
always forced the driver to assume the ATPX dGPU power down
was present, but this causes problems on boards where it is
not leading to GPU hangs when attempting to power down the
dGPU.  Manual dGPU power down is not currently supported in
the Linux driver.  Some laptops indicate that the ATPX
dGPU power down method is not present, but it actually
apparently is.  I'm not sure if this is a bios bug and it
should be set or if there is a reason it was unset and the
method should not be used.  The adds a quirk to forcibly enable
it on specific systems to restore the old behavior.  This is
not an issue on other OSes since both the ATPX and the manual
driver power down methods are supported.

bug:
https://bugzilla.kernel.org/show_bug.cgi?id=115321

Signed-off-by: Alex Deucher <alexander.deucher@amd.com>