drm/amd/powerplay: SCLK will have a big drop with low VDDC when PPlib was enabled.
cherry pick CL1262020 from PPLib.

SMC uses CurrSclkPllRange structure to keep track of what range of
PLL SCLK is sitting on.  Driver overwrites this value to 0 because
it's part of DPM table and driver doesn't program this.
This change will set this field to 0xFF every time there's a
init SMC table call.

Change-Id: Id6f12cc33d447c97dd7a5b914faf9a5f799f32e7
Signed-off-by: Rex Zhu <Rex.Zhu@amd.com>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>