drm/amdkfd: Unmapping static queues when a process terminates
In the original code, the static queues were also unmapped when a
process terminates. However, when we tried to fix a multi-process
dispatch failure in SWDEV-89586, we excluded the static queues in
order not to reset the queues when unmapping. Now we include back
the static queues as the new interface of execute_queues_cpsch()
allows us to specify the queue filter.

Change-Id: I09959e96fc6ee2990ff80c251910dcc7c0c631c8
Signed-off-by: Yong Zhao <yong.zhao@amd.com>