drm/amdkfd: Enable user trap handler on non-CWSR ASICs
AMDKFD_IOC_SET_TRAP_HANDLER requires a context-save trap handler
to be installed, from which a jump to the user-provided trap handler
can be initiated. We would like to have this feature on ASICs which
do not support context-save/restore.

When a context-save handler is not installed configure the trap handler
registers to point directly to the user trap handler.

Also remove the legacy microcode version check since the KFD relies on
functionality present only in newer versions.

Signed-off-by: Jay Cornwall <Jay.Cornwall@amd.com>