drm/amdkfd: Fix the PTE not cleared on BO unmapping bug
This bug hides very well because the BO mapping and unmapping are
working in charm to not erase the PTE and not release the vm mapping
until the BO gets freed. This is also the case for memory eviction.

We fix the bug by correctly erase the PTE and release the vm mapping
on unmapping and rewrite the PTE and recreate vm mapping on mapping.

Fix BUG: SWDEV-89590 SWDEV-106528

Change-Id: If7312e8912aedc9365f359ec99719bf76d493665
Signed-off-by: Yong Zhao <yong.zhao@amd.com>