drm/amd/hdcp_service: add hdcp security service component
HDCP securty service plays an important role during
the authentication phase of the HDCP receiver with
the HDCP transmitter. The HDCP transmitter and the
receiver exchange a series of security messages to
establish the authenticity of the receiver. The
receiver that is successfully authenticated by the
transmitter is authorized to play the protected
content that is sent by the transmitter.

HDCP security service offloads security critical tasks
to HDCP TA running as a secure application on PSP.

HDCP security service uses interface provided by DAL
to read and write display registers over I2C/AUX.

The HDCP authentication sequence is managed by HDCPSS.

Change-Id: Ia01e8ccd6607e7d4c686e5a4e9ca906eccc57c63
Signed-off-by: Maulik Mankad <maulik.mankad@amd.com>
Signed-off-by: Sagar Alatgi <Sagar-mahesh.Alatgi@amd.com>