drm/amdgpu: Fix userptr restore race condition with free
Update_user_pages needs to drop the lock. This could lead to freeing
a userptr BO while a restore was in progress. In that case
cancel_restore_locked and restore_mem_worker would both call mmput
and lead to corruption of kernel data structures.

Fix this by marking the mem object as busy while dropping the lock
for user page udpates. When canceling restore, wait for busy mem
objects. This ensures that a restore that has already started will
be able complete before the BO gets freed. When the restore completes
it sets mem->mm to NULL and calls mmput. Thes cancel_restore_locked
will not call mmput again.

Bug: SWDEV-112697
Change-Id: I6c76f7559957992303cf69b7a26360886f090990
Signed-off-by: Felix Kuehling <Felix.Kuehling@amd.com>