drm/amdkfd: Remove BUG_ONs
FOr the most part, these are internal functions and we don't need to
check the validity of the parameters. A BUG_ON would be essentially
the same as a NULL reference anyways. Remove most of the BUG_ONs for
simple parameter checking, and change the ones that matter to either
WARN_ON (for void functions) or pr_err (for int functions)

Change-Id: I7268b3c29bac5e87545a1a18ae5acdec8175b684
Signed-off-by: Kent Russell <kent.russell@amd.com>

 Conflicts:
	drivers/gpu/drm/amd/amdgpu/amdgpu_amdkfd.c
	drivers/gpu/drm/amd/amdkfd/kfd_device_queue_manager.c