drm/amdgpu: Allow get_user_pages to fail in restore worker
Avoid stalling queues indefinitely if get_user_pages fails. This type
of failure is indicative of a user mode bug where memory is freed that
is still mapped to the GPU. Let the restore continue and update the
page table with invalid entries for the failed BO. If the GPU tries to
access it, the application will crash with a VM fault.

Bug: SWDEV-117987
Change-Id: Ica23b8d562e8268d80109192f1c5f0c16eb72de0
Signed-off-by: Felix Kuehling <Felix.Kuehling@amd.com>