amdgpu: Use vm_context instead of is_kfd_vm
We really only care about the number of KFD VMs as we want to ensure
that the power profile is set to Compute if we have at least one VM that
is KFD-based (compute-based). Instead of a boolean, use vm_context and
store this in the vm object for reference. This also allows us to merge
smoothly with amdgpu after Harish's upstream changes

Change-Id: I52cc50d8c41d1b4404450cfbddcdec129442affd
Signed-off-by: Kent Russell <kent.russell@amd.com>

 Conflicts:
	drivers/gpu/drm/amd/amdgpu/amdgpu_vm.c