drm/amdkfd: Change process lock back to mutex
This was changed to a read/write semaphore due to complicated locking
dependencies in userptr eviction scenarios. The same problem was
solved better by adding a reference count to struct process instead
of misusing a lock to control the process lifetime. This removed
the problematic lock dependencies.

Mutexes are better optimized than semaphores. Only very few cases of
read locks were left in the code, so mutex should be better performance
overall.

This also reduces the amount of change we need to upstream and justify
in the upstream reviews.

Change-Id: I962e72dd1e870f182303272339d12b9301506f00
Signed-off-by: Felix Kuehling <Felix.Kuehling@amd.com>